requirements.txt
accelerate
aitextgen
arrange
bitsandbytes
google-generativeai
huggingface_hub
ipywidgets
langchain
langchain_google_genai
numpy
openai
pandas
pdf2image
pdfminer
pinecone-client
pypdf
python-dotenv
sentence_transformers
semanticscholar
streamlit
tiktoken
torchtext
torchviz
transformers
unstructured
rouge-score
rouge
python-magic
opencv-python
pillow-heif
pdfminer.six
accelerate
tensorboardX
peft
trl
tiktoken
seaborn
weaviate-client
bert_score
evaluate
wandb
GitPython #clone the repository
# pip install --upgrade jupyter ipywidgets
chromadb
llama-index
llama-index-llms-huggingface
llama-index-embeddings-huggingface
llama-index-embeddings-instructor
llama-index-vector-stores-weaviate
llama-index-vector-stores-lancedb
llama-index-multi-modal-llms-openai
llama-index-embeddings-clip
llama-index-readers-file
ftfy 
regex 
tqdm
openai-whisper
matplotlib
scikit-image
lancedb
moviepy
pytube
pydub
SpeechRecognition
ffmpeg-python
soundfile
git+https://github.com/openai/CLIP.git
ragas
chainlit
python-magic-bin
trl
py7zr
langchain_groq
langchain_core
auto-gptq
optimum
langserve[all]
sse_starlette
langchain-huggingface
git+https://github.com/huggingface/transformers